Nueva entidad Oferta y corrección de errores

+++ src/main/java/com/example/negocio/dto/compra/CompraFullDTO.java
@@ -4,13 +4,15 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 import java.util.List;
 
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 public class CompraFullDTO {
     private Long idCompra;
-    private Double total;
+    private BigDecimal total;
     private LocalDateTime fechaHora;
     private String proveedor;
     private String usuario;
