+++ src/main/java/com/example/negocio/dto/compra/DetalleCompraFullDTO.java
@@ -5,9 +5,11 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 public class DetalleCompraFullDTO {
     private String producto;
     private Integer cantidad;
-    private Double costoUnitario;
+    private BigDecimal costoUnitario;
 }
