+++ src/main/java/com/example/negocio/dto/producto/ProductoDTO.java
@@ -5,11 +5,13 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 public class ProductoDTO {
     private String nombre;
-    private Double precio;
-    private Double costo;
+    private BigDecimal precio;
+    private BigDecimal costo;
     private Integer stock;
     private Integer stockMinimo;
     private Long idMarca;
