+++ src/main/java/com/example/negocio/dto/promocion/PromocionAbmDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
@@ -11,7 +12,7 @@ public class PromocionAbmDTO {
     private Long idPromocion;
     private String nombre;
     private String descripcion;
-    private Double precio;
+    private BigDecimal precio;
     private Boolean estado;
     private List<DetallePromocionAbmDTO> detalles;
 }
