+++ src/main/java/com/example/negocio/dto/promocion/PromocionListaDTO.java
@@ -4,10 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 public class PromocionListaDTO {
     private Long idPromocion;
     private String nombre;
-    private Double precio;
+    private BigDecimal precio;
 }
