+++ src/main/java/com/example/negocio/dto/venta/DetalleVentaListaDTO.java
@@ -4,11 +4,12 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 public class DetalleVentaListaDTO {
     private String producto;
     private String promocion;
     private Integer cantidad;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 }
