+++ src/main/java/com/example/negocio/entity/DetalleVenta.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
 @Entity @Table
 public class DetalleVenta {
@@ -13,7 +15,7 @@ public class DetalleVenta {
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long idDetalleVenta;
     private Integer cantidad;
-    private Double precioUnitario;
+    private BigDecimal precioUnitario;
 
     @ManyToOne @JoinColumn(name = "idProducto")
     private Producto producto;
