+++ src/main/java/com/example/negocio/entity/Promocion.java
@@ -5,6 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data @Builder @AllArgsConstructor @NoArgsConstructor
@@ -15,7 +17,7 @@ public class Promocion {
     private Long idPromocion;
     private String nombre;
     private String descripcion;
-    private Double precio;
+    private BigDecimal precio;
     private Boolean estado;
 
     @OneToMany(mappedBy = "promocion", cascade = CascadeType.ALL, orphanRemoval = true)
